makes core private, moves mocks to conftest
